+++ src/components/physics/properties.rs
@@ -5,26 +5,26 @@ use amethyst::ecs::prelude::{Component, VecStorage};
 #[derive(Debug, Default, Clone)]
 pub struct PhysicalProperties {
     /// The weight of the entity itself, like a car's empty mass.
-    pub mass: f64,
+    pub mass: f32,
     /// The resistance against rotational acceleration.
     /// If it is `None`, this entity cannot be rotated.
-    pub inertia: Option<f64,>,
+    pub inertia: Option<f32,>,
     /// The value of the friction coefficient.
     /// If it is `None`, this entity has supposedly no ground contact,
     /// e.g. the Ship, as it is supported by the Tracks.
-    pub friction: Option<f64,>,
+    pub friction: Option<f32,>,
     /// The value of the air-resistance coefficient.
     /// If it is `None`, this entity has supposedly no air resistance,
     /// e.g. the Drill, as it is supposed to be pulled into the Ship when flying.
-    pub air_resistance: Option<f64,>,
+    pub air_resistance: Option<f32,>,
 }
 
 impl PhysicalProperties {
     pub fn new(
-        mass: f64,
-        inertia: Option<f64,>,
-        friction: Option<f64,>,
-        air_resistance: Option<f64,>,
+        mass: f32,
+        inertia: Option<f32,>,
+        friction: Option<f32,>,
+        air_resistance: Option<f32,>,
     ) -> Self {
         PhysicalProperties {
             mass,
